requirements/base.txt
# requirements/base.txt
pyside6>=6.0.0
anthropic>=0.3.0
openai>=1.0.0
python-dotenv>=0.19.0
rich>=10.0.0
typer>=0.4.0
pyyaml>=6.0.0
psutil>=5.8.0
watchdog>=2.1.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.3.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0